# elza-web/src/main/resources/frontend/swagger.yaml
swagger: '2.0'
info:
  title: Elza
  description: ElzaAPI
  version: "9.0.0"
# the domain of the service
host: localhost:8080
schemes:
  - http
# will be prefixed to all paths
basePath: /api/
produces:
  - application/json
consumes:
  - application/json
paths:
...
  /registry:
    get:
      summary: Find registry by params
      parameters:
        - name: search
          in: query
          description: ID
          required: false
          type: string
        - name: from
          in: query
          description: Offset of results
          required: true
          type: integer
        - name: count
          in: query
          description: Count of results
          required: true
          type: integer
        - name: registerTypeId
          in: query
          description: register type ID
          required: false
          type: integer
        - name: parentRecordId
          in: query
          description: Parent record id
          required: false
          type: integer
        - name: versionId
          in: query
          description: version ID
          required: false
          type: integer
      tags:
        - Registry
      responses:
        200:
          description: Registry
          schema:
            type: object
            properties:
              count:
                type: integer
              rows:
                type: array
                items:
                  $ref: '#/definitions/RegRecordSimple'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Get party by ID
      parameters:
        - name: party
          in: body
          description: Party
          required: true
          schema:
            $ref: '#/definitions/ParPartyVO'
      tags:
        - Party
      responses:
        200:
          description: Created object
          schema:
            $ref: '#/definitions/ParPartyVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /registry/{recordId}:
    get:
      summary: Get record by ID
      parameters:
        - name: recordId
          in: path
          description: ID
          required: true
          type: integer
      tags:
        - Registry
      responses:
        200:
          description: Record
          schema:
            $ref: '#/definitions/RegRecordVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: Update record by ID
      parameters:
        - name: recordId
          in: path
          description: ID
          required: true
          type: integer
        - name: record
          in: body
          description: Record
          required: true
          schema:
            $ref: '#/definitions/RegRecordVO'
      tags:
        - Registry
      responses:
        200:
          description: Record
          schema:
            $ref: '#/definitions/RegRecordVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete record by ID
      parameters:
        - name: recordId
          in: path
          description: ID
          required: true
          type: integer
      tags:
        - Registry
      responses:
        200:
          description: Successful delete
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /registry/findRecordForRelation:
    get:
      summary: Find registry by params for relation
      parameters:
        - name: search
          in: query
          description: ID
          required: false
          type: string
        - name: from
          in: query
          description: Offset of results
          required: true
          type: integer
        - name: count
          in: query
          description: Count of results
          required: true
          type: integer
        - name: roleTypeId
          in: query
          description: role type ID
          required: true
          type: integer
        - name: partyId
          in: query
          description: Party ID
          required: true
          type: integer
      tags:
        - Registry
      responses:
        200:
          description: Registry
          schema:
            type: object
            properties:
              count:
                type: integer
              rows:
                type: array
                items:
                  $ref: '#/definitions/RegRecordSimple'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /scopes:
    get:
      summary: get all scopes
      tags:
        - Registry
      responses:
        200:
          description: Scopes
          schema:
            type: array
            items:
              $ref: '#/definitions/RegScopeVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Create scope
      parameters:
        - name: scope
          in: body
          description: Scope
          required: true
          schema:
            $ref: '#/definitions/RegScopeVO'
      tags:
        - Party
      responses:
        200:
          description: Created object
          schema:
            $ref: '#/definitions/RegScopeVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /scopes/{scopeId}:
    put:
      summary: Update scope by ID
      parameters:
        - name: scopeId
          in: path
          description: ID
          required: true
          type: integer
        - name: record
          in: body
          description: Scope
          required: true
          schema:
            $ref: '#/definitions/RegScopeVO'
      tags:
        - Registry
      responses:
        200:
          description: Scope
          schema:
            $ref: '#/definitions/RegScopeVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete scope by ID
      parameters:
        - name: scopeId
          in: path
          description: ID
          required: true
          type: integer
      tags:
        - Registry
      responses:
        200:
          description: Successful delete
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /fundScopes:
    get:
      summary: get scopes for fund version or from yaml
      parameters:
        - name: versionId
          in: query
          description: version Id
          required: false
          type: integer
      tags:
        - Registry
      responses:
        200:
          description: Scopes
          schema:
            type: array
            items:
              $ref: '#/definitions/RegScopeVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /registry/recordTypes:
    get:
      summary: Record Types
      tags:
        - Registry
      responses:
        200:
          description: Record
          schema:
            type: array
            items:
              $ref: '#/definitions/RegRegisterTypeVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /registry/recordTypesForPartyType:
    get:
      summary: Record Types for party types
      tags:
        - Registry
      parameters:
        - name: partyTypeId
          in: query
          description: ID
          required: false
          type: integer
      responses:
        200:
          description: Record
          schema:
            type: array
            items:
              $ref: '#/definitions/RegRegisterTypeVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /registry/variantRecord:
    post:
      summary: Create Variant record
      parameters:
        - name: variant record
          in: body
          description: Variant record
          required: true
          schema:
            $ref: '#/definitions/RegVariantRecordVO'
      tags:
        - Registry
      responses:
        200:
          description: Created object
          schema:
            $ref: '#/definitions/RegVariantRecordVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /registry/variantRecord/{variantRecordId}:
    put:
      summary: Update variant record by ID
      parameters:
        - name: variantRecordId
          in: path
          description: ID
          required: true
          type: integer
        - name: variantRecord
          in: body
          description: Record
          required: true
          schema:
            $ref: '#/definitions/RegVariantRecordVO'
      tags:
        - Registry
      responses:
        200:
          description: Record
          schema:
            $ref: '#/definitions/RegVariantRecordVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete variant record by ID
      parameters:
        - name: variantRecordId
          in: path
          description: ID
          required: true
          type: integer
      tags:
        - Registry
      responses:
        200:
          description: Successful delete
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /registry/regCoordinates:
    post:
      summary: Create coordinates
      parameters:
        - name: variant record
          in: body
          description: coordinates
          required: true
          schema:
            $ref: '#/definitions/RegCoordinatesVO'
      tags:
        - Registry
      responses:
        200:
          description: Created object
          schema:
            $ref: '#/definitions/RegCoordinatesVO'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  ParPartyTypeVO:
    type: object
    properties:
      id:
        type: integer
      code:
        type: string
      name:
        type: string
      description:
        type: string
  ParPartyVO:
    type: object
    properties:
      id:
        type: integer
      partyType:
        $ref: '#/definitions/ParPartyTypeVO'
      history:
        type: string
      sourceInformation:
        type: string
      relations:
        type: string
      partyNames:
        type: string
      creators:
        type: array
        items:
          type: object
      record:
        $ref: '#/definitions/ParPartyTypeVO'
      characteristics:
        type: string
      version:
        type: integer
  RegVariantRecordVO:
    type: object
  RegRecordSimple:
    type: object
  RegRecordVO:
    type: object
  RegCoordinatesVO:
    type: object
  RegRegisterTypeVO:
    type: object
  RegScopeVO:
    type: object
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string